Ensure we get the correct document title:
If the title includes other tags or the h3 tag includes other
parameters, the comparison with 't' will not include the text. The
change makes sure we are comparing against the text.

Also, after checking enough pages, the document title always contains
the word LEctionary, so using that to find it.